==> 2. Cleaning and Preprocessing/2. sentences_preprocessing.ipynb <==
# ATTN: This script should be run AFTER language detection is completed.

# Drop non-English job descriptions

### START HERE IF SOURCING FROM DF_JOBS_RAW_LANGUAGE_DETECTED
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_raw_language_detected.pkl')


# len = 62577
df_jobs.info()


# nl = 44863, en = 17591, ['en', 'nl'] = 8, ['nl', 'en'] = 9
df_jobs['Language'].value_counts()


%%time
# The majority of ['en', 'nl'] labeled job descriptions contain mostly English
df_jobs['Language'] = df_jobs['Language'].progress_apply(
    lambda lang: 'en' if lang == "['en', 'nl']" else lang
)


# Drop non-English ads
df_jobs = df_jobs.drop(
    df_jobs[
        df_jobs['Language'] != 'en'
    ].index,
        axis='index',
)


# 17599
len(df_jobs)


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_english_only.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_english_only.csv', index=False)


# Fix abbreviations in job descriptions

### START HERE IF SOURCING FROM DF_JOBS_RAW_ENGLISH_ONLY
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


# Function to make a list of punctuations that determine sentence boundry, i.e., split characters
def make_custom_punct_chars(main_punct_chars=None, repeated_punct_chars=None):
    if main_punct_chars is None:
        main_punct_chars = [':', '|']
    if repeated_punct_chars is None:
        repeated_punct_chars = ['\n', ',']
    custom_punct_chars = []
    temp_multi = []
    temp_spaced = []

    for punct_char in main_punct_chars:
        custom_punct_chars+= f'{punct_char}', f'{punct_char} '

    for idx in range(4):
        for punct_char in repeated_punct_chars:
            temp_multi.append(f'{punct_char}'*int(idx+1))
            temp_spaced.append(f'{punct_char} '*int(idx+1))

    for multi, spaced in zip(temp_multi, temp_spaced):
        custom_punct_chars+= multi, spaced

    custom_punct_chars.remove(',')
    custom_punct_chars.remove(', ')

    return custom_punct_chars


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_raw_english_only.pkl')


# len = 17599
df_jobs.info()


# Conver Job ID and Sentence to str
str_cols = [
    'Job ID',
    'Job Description',
]

for col in str_cols:
    df_jobs[col] = df_jobs[col].astype(str)
    df_jobs[col] = df_jobs[col].progress_apply(lambda x: x.strip())
    df_jobs[col] = df_jobs[col].progress_apply(lambda x: unicodedata.normalize('NFKD', x.encode('ascii', 'ignore').decode('utf-8', 'ignore')))
    print(f'{col} converted to str.' if all(df_jobs[col].progress_apply(lambda x: isinstance(x, str))) else f'{col} NOT converted to str.')


df_jobs['Job Description'].head()


# count=17599, unique=17594, top=, freq=1
df_jobs['Job Description'].describe()


df_jobs = df_jobs.drop_duplicates(subset=['Job Description'], keep='first')


# len=17594
len(df_jobs)


df_jobs['Job Description'].head()


# count=17594, unique=17594, top=, freq=1
df_jobs['Job Description'].describe()


custom_punct_chars = make_custom_punct_chars()
print(custom_punct_chars)


# Drop na, empty strings, punctuations, and keep only sentences of length more than 0
with open(f'{data_dir}punctuations.txt', 'rb') as f:
    custom_punct_chars = pickle.load(f)

df_jobs = df_jobs.dropna(subset=['Job Description'])
df_jobs = df_jobs.loc[
    (df_jobs['Job Description'].isnull() == False)
    & (df_jobs['Job Description'].isna() == False)
    & (df_jobs['Job Description'].notnull() == True)
    & (df_jobs['Job Description'].notna() == True)
    & (df_jobs['Job Description'].astype(bool) == True)
    & (df_jobs['Job Description'].progress_apply(len) > 0)
    & (pd.isna(df_jobs['Job Description']) == False)
    & (df_jobs['Job Description'].astype(str).isin(custom_punct_chars) == False)
    & (df_jobs['Job Description'].astype(str).isin(non_whitespace_nan_list) == False)
    & (df_jobs['Job Description'].astype(str).isin(list(string.punctuation)) == False)
    & (df_jobs['Job Description'].astype(str).isin(list(string.whitespace)) == False)
]
df_jobs = df_jobs.dropna(subset=['Job Description'])


df_jobs['Job Description'].head()


df_jobs['Job Description'].describe()


# Remove duplicate sentences from the same job ad
df_jobs = df_jobs.drop_duplicates(subset=['Job ID', 'Job Description'])


# Check row are string or invalid
for idx, row in df_jobs['Job Description'].items():
    if (not isinstance(row, str)) or (row in nan_list) or (row in custom_punct_chars) or (row in list(string.punctuation)) or (pd.isna(row) == True):
        print(idx, row)


# len=17593
len(df_jobs)


df_jobs['Job Description'].head()


# count=17593, unique=17593
df_jobs['Job Description'].describe()


df_jobs['Job Description'] = df_jobs.loc[df_jobs['Job Description'].notnull(), 'Job Description'].progress_apply(
    lambda job_description: ' '.join(job_description.split('/')) if '/' in job_description else job_description
)


abb_dict = {
    r'incl\.': 'including',
    r'e\.g\.': 'for example',
    r'e\.g': 'for example',
    r'etc\.': 'et cetera',
    r'\"': '',
}


%%time
df_jobs['Job Description'] = df_jobs['Job Description'].replace(abb_dict, regex=True)


df_jobs['Job Description'].head()


# count= 17593, unique= 17593
df_jobs['Job Description'].describe()


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_fixed_abbreviations.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_fixed_abbreviations.csv', index=False)


# Add English and Dutch language requirement columns

### START HERE IF SOURCING FROM DF_JOBS_RAW_FIXED_ABBREVIATIONS
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_raw_fixed_abbreviations.pkl')


df_jobs.info()


%%time
# Add language requirement column
# Use regex to find language requirement
lang_requirements = {
    'Dutch Requirement in Job Ad': dutch_requirement_pattern, 'English Requirement in Job Ad': english_requirement_pattern
}

for lang_req, lang_req_pattern in lang_requirements.items():

    if lang_req in df_jobs.columns:
        df_jobs = df_jobs.drop(columns=[lang_req])
    df_jobs[lang_req] = np.where(
        df_jobs['Job Description'].str.contains(lang_req_pattern),
        'Yes',
        'No',
    )
    df_jobs[lang_req] = df_jobs[lang_req].astype('category').cat.reorder_categories(['No', 'Yes'], ordered=True)
    df_jobs[lang_req] = pd.Categorical(df_jobs[lang_req], categories=['No', 'Yes'], ordered=True)

assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_language_requirement.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_english_requirement.csv', index=False)


# Yes = 298 job ads
df_jobs['Dutch Requirement in Job Ad'].value_counts()


# Yes = 545 job ads
df_jobs['English Requirement in Job Ad'].value_counts()


# Dummy coding for language requirement
df_jobs = df_jobs.join(pd.get_dummies(df_jobs[['Dutch Requirement in Job Ad', 'English Requirement in Job Ad']], dtype='int64'))


df_jobs['Dutch Requirement in Job Ad_Yes'].value_counts()


df_jobs['English Requirement in Job Ad_Yes'].value_counts()


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_language_requirement.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_language_requirement.csv', index=False)


# Add data from Sectors dataframe (see CBS directory under '1. Scraping' directory)


### START HERE IF SOURCING FROM DF_JOBS_RAW_LANGUAGE_REQUIREMENT
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_raw_language_requirement.pkl')


df_jobs.info()


try:
    df_sectors_all = pd.read_pickle(f'{table_save_path}Sectors Output from script.pkl')
except FileNotFoundError:
    cbs_notebook = '\\'.join(f'{scraped_data}CBS/CBS.ipynb')
    %run $cbs_notebook import df_sectors_all # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_sectors_all.columns


df_sectors_all.columns = [
    '_'.join(col)
    if 'SBI Sector Titles' not in col
    and 'Total Workforce' not in col
    and 'Sector Titles' not in col
    else col[-1]
    for col in df_sectors_all.columns
]


df_sectors_all = df_sectors_all.rename(
    columns={
        'Keywords': 'Search Keyword',
        'Keywords n': 'Keywords Count',
        'Code': 'Sector Code',
        'Sector Name': 'Sector',
        'SBI 2008 Sector Name': 'Sector',
        'Gender_Sectoral Gender Segregation_Dominant Category': 'Gender',
        'Age_Sectoral Age Segregation_Dominant Category': 'Age',
        'n': 'Sector_n',
    },
)
df_sectors_all = df_sectors_all.rename(columns={element: re.sub(r' \(\W*45 years\)', '', element) for element in df_sectors_all.columns.tolist()})


df_sectors_all.columns


df_sectors_all = df_sectors_all.explode(
    'Search Keyword', ignore_index=True
)


df_sectors_all.columns


# len 101
len(df_sectors_all)


#### Before adding sector data, make sure keywords are correct as to not have any missing sector data when merging

# This is a manually collected dictionary of incorrect/faulty keywords in scraped site data
with open(f'{scraped_data}CBS/Data/keyword_trans_dict.txt') as f:
    keyword_trans_dict = json.load(f)

def fix_keywords(df_temp):

    if len(df_temp) > 0 and isinstance(df_temp, pd.DataFrame):
        for key, value in keyword_trans_dict.items():
            df_temp.loc[
                df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(
                lambda x: x.lower().strip()
                ) == str(key).lower().strip(), 'Search Keyword'
            ] = str(value).lower().strip()

        unfixed = df_temp.loc[
            df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(lambda x: x.lower().strip()).isin([x.lower().strip() for x in list(keyword_trans_dict.keys())])
        ]

        if len(unfixed) != 0:
            for key, value in keyword_trans_dict.items():
                for idx, row in df_temp.iterrows():
                    if row['Search Keyword'].astype(str).lower().strip() == str(key).lower().strip():
                        df_temp.loc[idx, 'Search Keyword'] = str(value).lower().strip()

        unfixed = df_temp.loc[
                df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(lambda x: x.lower().strip()).isin([x.lower().strip() for x in list(keyword_trans_dict.keys())])
            ]
        if len(unfixed) != 0:
            print('Some keywords were not fixed. Please check file unfixed_keywords.txt in data directory.')
            with open(f'{data_dir}unfixed_keywords.txt', 'w') as f:
                json.dump(unfixed, f)

    return df_temp

# Fix keywords
if len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]) != 0:
    print('Some search keywords did not match a sector. Fixing')
#     print(set(df_jobs['Search Keyword'].loc[df_jobs['Sector'].isna()].to_list()))
    print(len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]))
    df_jobs = fix_keywords(df_jobs)
#     print(set(df_jobs['Search Keyword'].loc[df_jobs['Sector'].isna()].to_list()))
    print(len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]))


df_jobs = df_jobs.merge(df_sectors_all, on='Search Keyword', how='left')


# 17593
len(df_jobs)


df_jobs.info()


#### Check if there is any missing sector data in the merged dataframe

df_jobs['Sector'].isna().sum()


if df_jobs['Sector'].isna().sum() != 0:
    print('Some search keywords did not match a sector. Fixing')
    print(set(df_jobs['Search Keyword'].loc[df_jobs['Sector'].isna()].to_list()))
    print(len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]))
    df_jobs = fix_keywords(df_jobs)
    print(set(df_jobs['Search Keyword'].loc[df_jobs['Sector'].isna()].to_list()))
    print(len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]))


# Female=3955, Male=6974, Mixed Gender=6664
df_jobs['Gender'].value_counts()


# Older=3778, Younger=2351, Mixed Age=11464
df_jobs['Age'].value_counts()


if df_jobs['Sector'].isna().sum() == 0:
    assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
    df_jobs.to_pickle(f'{df_save_dir}df_jobs_including_sector_data.pkl')
    df_jobs.to_csv(f'{df_save_dir}df_jobs_including_sector_data.csv', index=False)

else:
    print(f"MISSING SECTOR DATA: COUNT {df_jobs['Sector'].isna().sum()}")


# Add interaction terms and categorical data


### START HERE IF SOURCING FROM DF_JOBS_INCLUDING_SECTOR_DATA
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_including_sector_data.pkl')


df_jobs.info()


df_jobs['Sector Job Advertisement Count'] = df_jobs['Sector'].map(df_jobs.groupby('Sector')['Job ID'].count().to_dict())


df_jobs['Sector Job Advertisement Count'].head()


df_jobs[['Sector', 'Sector Job Advertisement Count']].value_counts()


df_jobs['Sector Job Advertisement Count'].describe()


df_jobs['Sector Gender Designation Job Advertisement Count'] = df_jobs['Gender'].map(df_jobs.groupby('Gender')['Job ID'].count().to_dict())


df_jobs['Sector Gender Designation Job Advertisement Count'].head()


df_jobs[['Gender', 'Sector Gender Designation Job Advertisement Count']].value_counts()


df_jobs['Sector Gender Designation Job Advertisement Count'].describe()


df_jobs['Sector Age Designation Job Advertisement Count'] = df_jobs['Age'].map(df_jobs.groupby('Age')['Job ID'].count().to_dict())


df_jobs['Sector Age Designation Job Advertisement Count'].head()


df_jobs[['Age', 'Sector Age Designation Job Advertisement Count']].value_counts()


df_jobs['Sector Age Designation Job Advertisement Count'].describe()


df_jobs = df_jobs.join(pd.get_dummies(df_jobs[['Gender', 'Age']], dtype='int64'))
df_jobs = df_jobs.rename({'Gender_Mixed Gender': 'Gender_Mixed', 'Age_Mixed Age': 'Age_Mixed'}, axis='columns')


df_jobs.columns


df_jobs = categorize_df_gender_age(df_jobs)


df_jobs['Interaction_Female_Older_% per Sector'] = df_jobs['Gender_Female_% per Sector'] * df_jobs['Age_Older_% per Sector']
df_jobs['Interaction_Female_Younger_% per Sector'] = df_jobs['Gender_Female_% per Sector'] * df_jobs['Age_Younger_% per Sector']
df_jobs['Interaction_Male_Older_% per Sector'] = df_jobs['Gender_Male_% per Sector'] * df_jobs['Age_Older_% per Sector']
df_jobs['Interaction_Male_Younger_% per Sector'] = df_jobs['Gender_Male_% per Sector'] * df_jobs['Age_Younger_% per Sector']


df_jobs.columns


get_df_info(df_jobs, ivs_all=ivs_all)


get_df_info(df_jobs, ivs_all=dvs)


df_jobs['Gender'].value_counts()


df_jobs['Gender_Female'].value_counts()


df_jobs['Gender_Num'].value_counts()


df_jobs['Platform'].value_counts()


df_jobs['Platform'] = df_jobs['Platform'].astype('category').cat.reorder_categories(platform_order, ordered=True)
df_jobs['Platform'] = pd.Categorical(df_jobs['Platform'], categories=platform_order, ordered=True)
df_jobs['Platform_Num'] = pd.to_numeric(df_jobs['Platform'].cat.codes).astype('int64')
df_jobs = df_jobs.join(pd.get_dummies(df_jobs[['Platform']], dtype='int64'))


# LinkedIn=13818, Indeed=3198, Glassdoor=3198
df_jobs[['Platform_Num', 'Platform_LinkedIn', 'Platform_Indeed', 'Platform_Glassdoor']].value_counts()


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_including_sector_genage_data.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_including_sector_genage_data.csv', index=False)


# Use spacy to split job ads to sentences


### START HERE IF SOURCING FROM DF_JOBS_INCLUDING_SECTOR_GENAGE_DATA
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import *  # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_jobs_including_sector_genage_data.pkl')


df_jobs.info()


df_jobs['Job Description'] = df_jobs['Job Description'].progress_apply(lambda x: unicodedata.normalize('NFKD', x.encode('ascii', 'ignore').decode('utf-8', 'ignore')))


df_jobs['Job Description'].head()


%%time
# Add sentencizer to spacy pipe and set custom punctuations
with open(f'{data_dir}punctuations.txt', 'rb') as f:
    custom_punct_chars = pickle.load(f)

if 'sentencizer' not in nlp.pipe_names:
    sentencizer = nlp.add_pipe('sentencizer')
sentencizer.punct_chars.update(custom_punct_chars)

if all(custom_punct_char in sentencizer.punct_chars for custom_punct_char in custom_punct_chars):
    with open(f'{data_dir}punctuations.txt', 'wb') as f:
        pickle.dump(sentencizer.punct_chars, f)

with open(f'{data_dir}punctuations.txt', 'rb') as f:
    custom_punct_chars = pickle.load(f)


# Add special cases to spacy
special_cases_dict = {
    'incl.': [{65: 'incl', 67: 'including'}],
    'incl. ': [{65: 'incl', 67: 'including'}],
    '(incl.': [{65: 'incl', 67: 'including'}],
    'etc.': [{65: 'etc', 67: 'et cetera'}],
    'etc. ': [{65: 'etc', 67: 'et cetera'}],
    'e.g.': [{65: 'e.g', 67: 'for example'}],
    'e.g. ': [{65: 'e.g', 67: 'for example'}],
    'andor ': [{65: 'andor', 67: 'and or'}],
    'and/or ': [{65: 'andor', 67: 'and or'}],
}

nlp.tokenizer.rules.update(special_cases_dict)


df_jobs['Job Description'].head()


df_jobs['Job Description'].describe()


%%time
# Spacy sentencize
if all(custom_punct_char in sentencizer.punct_chars for custom_punct_char in custom_punct_chars):
    df_jobs['Job Description spacy_sentencized'] = df_jobs['Job Description'].progress_apply(
        lambda job_description: [
            sent
            for sentence in nlp(job_description).sents
            for sent in re.split(pattern, sentence.text)
            if sent is not None
            and bool(sent) == True
            and isinstance(sent, str)
            and sent != '...'
            and sent not in nan_list
            and sent not in custom_punct_chars
            and sent not in list(string.punctuation)
            and sent not in list(string.whitespace)
            ]
    )

assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_sentencized.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_sentencized.csv', index=False)


# len=17593
len(df_jobs)


df_jobs['Job Description spacy_sentencized'].head()


df_jobs['Job Description spacy_sentencized'].describe()


# Drop na, empty strings, punctuations, and keep only sentences of length more than 0
df_jobs = df_jobs.dropna(subset=['Job Description spacy_sentencized'])
df_jobs = df_jobs.loc[
    (df_jobs['Job Description spacy_sentencized'].isnull() == False)
    & (df_jobs['Job Description spacy_sentencized'].isna() == False)
    & (df_jobs['Job Description spacy_sentencized'].notnull() == True)
    & (df_jobs['Job Description spacy_sentencized'].notna() == True)
    & (df_jobs['Job Description spacy_sentencized'].astype(bool) == True)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).progress_apply(len) > 0)
    & (pd.isna(df_jobs['Job Description spacy_sentencized']) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(custom_punct_chars) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(non_whitespace_nan_list) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(list(string.punctuation)) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(list(string.whitespace)) == False)
]
df_jobs = df_jobs.dropna(subset=['Job Description spacy_sentencized'])


# len=17593
len(df_jobs)


df_jobs['Job Description spacy_sentencized'].head()


# count=17593, unique=17028
df_jobs['Job Description spacy_sentencized'].describe()


# sentences=1007
df_jobs['Job Description spacy_sentencized'].progress_apply(len).max()


df_jobs['Job Description spacy_sentencized'].loc[df_jobs['Job Description spacy_sentencized'].progress_apply(len).idxmin()]


# Remove sentences with only punctuations, whitespace, and nan, etc from list of sentences
with open(f'{data_dir}punctuations.txt', 'rb') as f:
    custom_punct_chars = pickle.load(f)

df_jobs['Job Description spacy_sentencized'] = df_jobs['Job Description spacy_sentencized'].progress_apply(
    lambda job_sentences_list: [
        sentence.strip()
        for sentence in job_sentences_list
        if sentence is not None
        and len(sentence) > 5
        and sentence not in nan_list
        and sentence not in custom_punct_chars
        and sentence not in list(string.punctuation)
        and sentence not in list(string.whitespace)
    ]
)


df_jobs['Job Description spacy_sentencized'].head()


# count=17593, unique=17000
df_jobs['Job Description spacy_sentencized'].describe()


# Drop na, empty strings, punctuations, and keep only sentences of length more than 0
df_jobs = df_jobs.dropna(subset=['Job Description spacy_sentencized'])
df_jobs = df_jobs.loc[
    (df_jobs['Job Description spacy_sentencized'].isnull() == False)
    & (df_jobs['Job Description spacy_sentencized'].isna() == False)
    & (df_jobs['Job Description spacy_sentencized'].notnull() == True)
    & (df_jobs['Job Description spacy_sentencized'].notna() == True)
    & (df_jobs['Job Description spacy_sentencized'].astype(bool) == True)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).progress_apply(len) > 0)
    & (pd.isna(df_jobs['Job Description spacy_sentencized']) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(custom_punct_chars) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(non_whitespace_nan_list) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(list(string.punctuation)) == False)
    & (df_jobs['Job Description spacy_sentencized'].astype(str).isin(list(string.whitespace)) == False)
]
df_jobs = df_jobs.dropna(subset=['Job Description spacy_sentencized'])


# len=17592
len(df_jobs)


df_jobs.info()


df_jobs['Job Description spacy_sentencized'].head()


# count=17592, unique=16999
df_jobs['Job Description spacy_sentencized'].describe()


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_sentencized.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_sentencized.csv', index=False)
